Bug in variable_array reverse iterator utilities, compiler warning in audio stream

+++ CAudioStream.h
@@ -955,7 +955,7 @@ namespace cpl
 					if (aSamples > 0)
 					{
 						// TODO: ensure aSamples < std::uint16_T::max()
-						frame.audioPacket = AudioFrame(AudioFrame::MessageType::AudioPacketSeparate, numChannels, static_cast<std::uint16_t>(aSamples * numChannels));
+						frame.audioPacket = AudioFrame(AudioFrame::MessageType::AudioPacketSeparate, static_cast<std::uint8_t>(numChannels), static_cast<std::uint16_t>(aSamples * numChannels));
 
 						auto const byteSize = static_cast<std::size_t>(aSamples * AudioFrame::element_size);
